backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
python-dotenv==1.0.1
pydantic-settings==2.5.2
python-multipart==0.0.9
alembic==1.13.2
# psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
pymysql==1.1.1
openpyxl==3.1.5
requests==2.32.3
